+++ sshd-core/src/main/java/org/apache/sshd/server/channel/ChannelSession.java
@@ -722,10 +722,10 @@ public class ChannelSession extends AbstractServerChannel {
         if (command instanceof AsyncCommandStreamsAware) {
             asyncOut = new ChannelAsyncOutputStream(
                     this, SshConstants.SSH_MSG_CHANNEL_DATA,
-                    isSendChunkIfRemoteWindowIsSmallerThanPacketSize());
+                    isSendChunkIfRemoteWindowIsSmallerThanPacketSize(SshConstants.SSH_MSG_CHANNEL_DATA));
             asyncErr = new ChannelAsyncOutputStream(
                     this, SshConstants.SSH_MSG_CHANNEL_EXTENDED_DATA,
-                    isSendChunkIfRemoteWindowIsSmallerThanPacketSize());
+                    isSendChunkIfRemoteWindowIsSmallerThanPacketSize(SshConstants.SSH_MSG_CHANNEL_EXTENDED_DATA));
             ((AsyncCommandStreamsAware) command).setIoOutputStream(asyncOut);
             ((AsyncCommandStreamsAware) command).setIoErrorStream(asyncErr);
         } else {
@@ -922,11 +922,22 @@ public class ChannelSession extends AbstractServerChannel {
     /**
      * Chance for specializations to vary chunking behaviour depending on the SFTP client version.
      *
-     * @return {@code true} if chunk data sent via {@link ChannelAsyncOutputStream} when reported remote window size is
-     *         less than its packet size
-     * @see    ChannelAsyncOutputStream#ChannelAsyncOutputStream(Channel, byte, boolean)
+     * @param  cmd                           Either {@link SshConstants#SSH_MSG_CHANNEL_DATA SSH_MSG_CHANNEL_DATA} or
+     *                                       {@link SshConstants#SSH_MSG_CHANNEL_EXTENDED_DATA
+     *                                       SSH_MSG_CHANNEL_EXTENDED_DATA} indicating the output stream type
+     * @return                               {@code true} if should chunk data sent via {@link ChannelAsyncOutputStream}
+     *                                       when reported remote window size is less than its packet size
+     * @see                                  ChannelAsyncOutputStream#ChannelAsyncOutputStream(Channel, byte, boolean)
+     * @throws UnsupportedOperationException if the command is neither of the supported ones
      */
-    protected boolean isSendChunkIfRemoteWindowIsSmallerThanPacketSize() {
-        return false;
+    protected boolean isSendChunkIfRemoteWindowIsSmallerThanPacketSize(byte cmd) {
+        if (cmd == SshConstants.SSH_MSG_CHANNEL_DATA) {
+            return CoreModuleProperties.ASYNC_SERVER_STDOUT_CHUNK_BELOW_WINDOW_SIZE.getRequired(this);
+        } else if (cmd == SshConstants.SSH_MSG_CHANNEL_EXTENDED_DATA) {
+            return CoreModuleProperties.ASYNC_SERVER_STDERR_CHUNK_BELOW_WINDOW_SIZE.getRequired(this);
+        } else {
+            throw new UnsupportedOperationException(
+                    "Unsupported channel data stream command: " + SshConstants.getCommandMessageName(cmd & 0xFF));
+        }
     }
 }